requirements.txt
actionlib==1.13.2
angles==1.9.13
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
bondpy==1.8.6
bosdyn-api==3.0.0
bosdyn-choreography-client==3.0.0
bosdyn-choreography-protos==3.0.0
bosdyn-client==3.0.0
bosdyn-core==3.0.0
bosdyn-mission==3.0.0
bottle==0.12.19
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==0.4.23
catkin-pkg-modules==0.4.23
catkin-tools==0.7.1
cbor==1.0.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.7
Click==7.0
colorama==0.4.4
conan==1.42.1
constantly==15.1.0
controller-manager==0.19.5
controller-manager-msgs==0.19.5
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.15.0
Cython==0.29.14
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.13
diagnostic-updater==1.10.4
distro==1.4.0
docopt==0.6.2
docutils==0.16
dynamic-reconfigure==1.7.1
empy==3.3.2
entrypoints==0.3
fasteners==0.16.3
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
grpcio==1.41.1
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==3.3
image-geometry==1.16.0
incremental==16.10.1
interactive-markers==1.12.0
Jinja2==2.11.3
joint-state-publisher==1.15.0
joint-state-publisher-gui==1.15.0
keyring==18.0.1
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.5.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
MarkupSafe==2.0.1
message-filters==1.15.13
moveit-commander==1.1.6
moveit-core==1.1.6
moveit-ros-planning-interface==1.1.6
moveit-ros-visualization==1.1.6
mpi4py==3.0.3
netifaces==0.10.4
node-semver==0.6.1
nose==1.3.7
numpy==1.17.4
oauthlib==3.1.0
osrf-pycommon==1.0.0
paramiko==2.6.0
patch-ng==1.17.4
pipreqs==0.4.11
pluginbase==1.0.1
protobuf==3.19.0
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyassimp==4.1.4
pycairo==1.16.2
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.2
Pygments==2.10.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-gnupg==0.4.5
python-qt-binding==0.4.4
python-snappy==0.5.3
PyTrie==0.2
pytz==2019.3
PyYAML==5.3.1
readchar==3.0.4
requests==2.26.0
requests-unixsocket==0.2.0
resource-retriever==1.12.6
roman==2.0.0
rosbag==1.15.13
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.21.0
rosdep-modules==0.21.0
rosdistro==0.8.3
rosdistro-modules==0.8.3
rosgraph==1.15.13
rosinstall==0.7.8
rosinstall-generator==0.1.22
roslaunch==1.15.13
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.13
rosmake==1.15.8
rosmaster==1.15.13
rosmsg==1.15.13
rosnode==1.15.13
rosparam==1.15.13
rospkg==1.3.0
rospkg-modules==1.3.0
rospy==1.15.13
rosservice==1.15.13
rostest==1.15.13
rostopic==1.15.13
rosunit==1.15.8
roswtf==1.15.13
rviz==1.14.10
SecretStorage==2.3.1
sensor-msgs==1.13.1
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smclib==1.8.6
soupsieve==1.9.5
srdfdom==0.6.2
systemd-python==234
tf==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
topic-tools==1.15.13
tqdm==4.62.3
Twisted==18.9.0
txaio==2.10.0
u-msgpack-python==2.1
urdfdom-py==0.4.6
urllib3==1.26.7
vcstools==0.1.42
wadllib==1.3.3
webencodings==0.5.1
wrapt==1.13.2
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.10
yarg==0.1.9
zope.interface==4.7.1
